@@ -1,106 +1,35 @@
 {
-  "name": "admin-devoo",
-  "version": "3.0.0",
-  "private": true,
-  "description": "VueJS version of Argon Dashboard 2 by Creative Tim",
-  "author": "Creative Tim",
+  "name": "EMIY-admin",
+  "version": "3.2.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "sitemap": "vue-cli-service sitemap"
-  },
-  "browser": {
-    "fs": false,
-    "path": false,
-    "os": false
-  },
-  "node": {
-    "fs": "empty"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview --port 4173",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "-": "0.0.1",
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@meforma/vue-toaster": "^1.3.0",
-    "@nuxtjs/robots": "^2.5.0",
-    "@nuxtjs/sitemap": "^2.4.0",
-    "@popperjs/core": "^2.10.2",
-    "@tinymce/tinymce-vue": "^4.0.7",
-    "autoprefixer": "10.4.5",
-    "axios": "^0.27.2",
-    "bootstrap-icons": "^1.8.3",
-    "bootstrap-vue": "^2.22.0",
-    "chart.js": "3.6.0",
-    "core-js": "3.6.5",
-    "dropzone": "6.0.0-beta.1",
-    "file-api": "^0.10.4",
-    "fs": "0.0.1-security",
-    "i": "^0.3.7",
-    "intl-tel-input": "^17.0.19",
-    "jenesius-vue-modal": "^1.5.3",
-    "jspdf": "^2.5.1",
+    "@mdi/js": "^7.0.96",
+    "@rushstack/eslint-patch": "^1.1.4",
+    "axios": "^0.27.0",
+    "chart.js": "^3.6.0",
+    "numeral": "^2.0.6",
     "jwt-decode": "^3.1.2",
-    "libphonenumber-js": "^1.10.13",
-    "maz-ui": "^3.0.4",
-    "npm": "^8.19.2",
-    "quill": "1.3.6",
-    "replace-in-file": "^6.3.5",
-    "sitemap": "^7.1.1",
-     
-    "v-easy-dialog": "^1.0.6",
-    "vue": "^3.2.36",
-    "vue-class-component": "^8.0.0-0",
-    "vue-cli-plugin-sitemap": "~2.3.0",
-    "vue-cookies": "^1.8.1",
-    "vue-ctk-date-time-picker": "^3.0.0-alpha.0",
-    "vue-final-modal": "^3.4.4",
-    "vue-flatpickr-component": "^9.0.5",
-    "vue-i18n": "^9.2.2",
-    "vue-iframes": "^0.0.19",
-    "vue-loading-button": "^0.2.0",
-    "vue-marquee-text-component": "^2.0.1",
-    "vue-meta": "^2.4.0",
-    "vue-native-dialog": "^0.0.3",
-    "vue-papa-parse": "^3.1.0",
-    "vue-phone-number-input": "^1.12.13",
-    "vue-router": "4.0.14",
-    "vue-svg-loader": "^0.16.0",
-    "vue-tinymce-editor": "^1.6.2",
-    "vue3-popper": "^1.5.0",
-    "vuejs3-datepicker": "^1.0.7",
-    "vuetify": "^3.0.1",
-    "vuex": "4.0.2"
+    "pinia": "^2.0.13",
+    "vue": "^3.3.4",
+    "vue-router": "^4.0.0"
   },
   "devDependencies": {
-    "@nuxtjs/vuetify": "^1.12.3",
-    "@types/fs-extra": "^9.0.13",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
-    "@vue/cli-plugin-babel": "^5.0.8",
-    "@vue/cli-plugin-eslint": "^5.0.8",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-service": "^5.0.8",
-    "@vue/compiler-sfc": "^3.2.37",
-    "@vue/eslint-config-prettier": "6.0.0",
-    "babel-eslint": "10.1.0",
-    "bootstrap": "^5.2.0-beta1",
-    "eslint": "^7.32.0",
-    "eslint-plugin-prettier": "3.3.1",
-    "eslint-plugin-vue": "7.0.0",
-    "jquery": "^3.6.0",
-    "papaparse": "^5.3.2",
-    "popper.js": "^1.16.1",
-    "prettier": "2.2.1",
-    "sass": "1.43.3",
-    "sass-loader": "10.1.1"
-  },
-  "bugs": {
-    "url": "https://github.com/creativetimofficial/vue-argon-dashboard/issues"
-  },
-  "homepage": "https://demos.creative-tim.com/vue-argon-dashboard/",
-  "license": "SEE LICENSE IN <https://www.creative-tim.com/license>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/creativetimofficial/vue-argon-dashboard.git"
+    "@rushstack/eslint-patch": "^1.1.4",
+    "@tailwindcss/forms": "^0.5.1",
+    "@vitejs/plugin-vue": "^3.0.3",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "autoprefixer": "^10.2.6",
+    "eslint": "^8.5.0",
+    "eslint-plugin-vue": "^9.0.0",
+    "postcss": "^8.3.0",
+    "postcss-import": "^14.0.2",
+    "prettier": "^2.7.1",
+    "tailwindcss": "^3.3.0",
+    "vite": "^3.0.9"
   }
 }
