@@ -1,4 +1,4 @@
-import VueCookies from 'vue-cookies'
+// import VueCookies from 'vue-cookies'
 
 import { api } from './axios';
 import jwt_decode from 'jwt-decode';
@@ -7,26 +7,26 @@ import jwt_decode from 'jwt-decode';
 const refreshAuthenticated = async () => {
   let decode = [];
   
-  if (
-    VueCookies.get('refreshToken') != undefined &&
-    VueCookies.get('refreshToken') != 'undefined'
-  ) {  
-     console.log('ici  *******',    VueCookies.get('refreshToken') );
+  // if (
+  //   VueCookies.get('refreshToken') != undefined &&
+  //   VueCookies.get('refreshToken') != 'undefined'
+  // ) {  
+  //    console.log('ici  *******',    VueCookies.get('refreshToken') );
     await api
       .post('/api/token/refresh', {
-        refreshToken: VueCookies.get('refreshToken'),
+        refreshToken:" VueCookies.get('refreshToken')",
       })
       .then((response) => {
-       console.log('nous effectuons le refresh token', response.data);
-        api.defaults.headers.common['Authorization'] =
-          'Bearer ' + String(response.data.token);
-        VueCookies.set('token', String(response.data.token));
-        VueCookies.set('refreshToken', String(response.data.refreshToken));
-        decode = jwt_decode(String(response.data.token));
-        VueCookies.set('nom', String(decode.nom));
-        VueCookies.set('prenom', String(decode.prenom));
-        VueCookies.set('phone', String(decode.phone));
-        VueCookies.set('id', String(decode.id));
+      //  console.log('nous effectuons le refresh token', response.data);
+      //   api.defaults.headers.common['Authorization'] =
+      //     'Bearer ' + String(response.data.token);
+      //   VueCookies.set('token', String(response.data.token));
+      //   VueCookies.set('refreshToken', String(response.data.refreshToken));
+      //   decode = jwt_decode(String(response.data.token));
+      //   VueCookies.set('nom', String(decode.nom));
+      //   VueCookies.set('prenom', String(decode.prenom));
+      //   VueCookies.set('phone', String(decode.phone));
+      //   VueCookies.set('id', String(decode.id));
     
         // console.log(decode.roles[1] == 'ROLE_ADMIN');
     
@@ -45,10 +45,10 @@ const refreshAuthenticated = async () => {
   //  router.push({ name: 'Signin' });
         }
       });
-  }
-  else {
-          // router.push({ name: 'Signin' });
-  }
+   // }
+  // else {
+  //         // router.push({ name: 'Signin' });
+  // }
 };
  
 
